requirements.txt
# Install with Python 3.10–3.13 (CrewAI is not available for 3.14+ on PyPI yet).
crewai>=0.76.0
python-dotenv>=1.0.0
